refactor(reasoning): WS-3 Phase 3 — extract makeObservationResult to kernel/utils (cycle 2 closure)
Move pure helper `makeObservationResult` from
`kernel/capabilities/act/tool-execution.ts` to
`kernel/utils/observation-helpers.ts`.

Closes the makeObservationResult cross-capability edge of cycle 2
(act ↔ reason). Reason capability (think.ts, think-guards.ts) and tests
now construct ObservationResult values via kernel/utils/ substrate, not
by importing from act/.

Revised from the thin spec's Tag-based DI proposal: pre-flight audit
showed the cross-cap inbound from reason→act is the pure helper, not
the effectful primitives (executeToolCall, executeNativeToolCall),
which remain in act/ as canonical owners. Tag overhead is unnecessary
for pure helpers — same pattern as Phase 1's tool-parsing extraction.

Scope:
- Create kernel/utils/observation-helpers.ts (pure, no Effect deps)
- Remove function definition + observation.ts symbol imports from
  tool-execution.ts; import makeObservationResult back from utils/ for
  the 6 internal call sites
- Update 2 src consumers: act.ts (split import), think-guards.ts,
  think.ts
- Update 3 test consumers: think-guards.test.ts, tool-execution.test.ts
  (split import), observation-trust-level.test.ts

Out of scope (per MissionBrief):
- Remaining reason→act edges via tool-gating / tool-parsing imports
  (think.ts, think-guards.ts) — Phase 4 territory
- executeToolCall / executeNativeToolCall Tag exposure — no cross-cap
  consumer needs it today
- truncateForDisplay move — not in cycle 2

Gates:
- Build: 38/38 green
- bun test: 5750 pass / 0 fail / 23 skip / 5773 total
- Typecheck baseline preserved (testing#cf-24, reasoning pre-existing
  arbitrator.budget.test/reactive-observer errors — both unchanged from
  branch baseline before this commit, verified via stash + retest)
